Debounce onSearchInput
Currently while the user types input into the search bar the search
event fires for every keypress. If they type the word "Hello" this means
that 4 wasted searches are performed and displayed for a fraction of a
second until the 5th and final search is displayed.

In a system with hundreds of tabs this greatly improves the
responsiveness.